Stop promtail service on user home change
When the promtail users home changes, the processes of the user must be
stopped. This checks if the promtail user already exists, get its
prameters and also checks if there is a promtail service unit file.
Before changing the promtail users home, the promtail service will be
stopped and restarted on change.